move ItemBankConfig to `/config/` directory
